requirements.txt
cycler==0.11.0
fonttools==4.33.3
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.0
matplotlib==3.5.2
numba==0.55.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pynndescent==0.5.6
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tqdm==4.64.0
umap-learn==0.5.3
